requirements.txt
biopython>=1.81
streamlit>=1.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0
langchain>=0.3.26,<1.0.0
langgraph>=0.0.40
langchain-core>=0.3.70,<1.0.0
pinecone>=7.0.0
sentence-transformers>=2.2.0,<2.3.0
transformers>=4.21.0,<5.0.0
huggingface-hub>=0.33.4,<1.0.0